KAFKAC-106 Support assigning document ID from multiple fields
Motivation
==========
The current `couchbase.document.id` config option assigns the document
ID from a single field of the document. Users want to combine multiple
fields and customize the formatting.

Modifications
=============
Change `couchbase.document.id` to be a format string with a placeholder
for each field to include in the ID.

For backwards compatibility, if the format string has no placeholders,
treat it like a single JSON pointer (the old way of doing things).

When removing the field from the document, just overwrite it with
spaces instead of creating a composite ByteBuf. This makes it easier
to remove multiple fields, and is more efficient than the old strategy.

Result
======
Document ID can be assigned from multiple fields joined by arbitrary
strings.

Change-Id: I878170d0993de97b3c8b871c2e2255fd8fc867a1
Reviewed-on: http://review.couchbase.org/95342
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>