KAFKAC-118 Source: vBucketUuid may be incorrect KAFKAC-161 Should include vBucket UUID in saved offset
Remove snapshot tracking since it has no effect on the correctness of
the Kafka stream and would blow out the heap during initial backfill
where the first snapshot contains all of history.

Removing snapshot tracking also makes it easier to accurately record
the vBucket UUID for each DCP event.

Store the vBucket UUID in the source offset. Use it to resume streaming.
This eliminates one possible cause of spurious rollback to zero when
the connector is restarted.

If rollback recovery fails, report the failure by submitting the error
to the fatal error queue. This causes the connector to terminate
instead of failing silently.

Change-Id: Ie13037961cba89d5be002dd8c0ff21fb7c23adbe
Reviewed-on: http://review.couchbase.org/119741
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>