KAFKAC-177 Source: Make RawJsonSourceHandler the default
Motivation
----------
One of the most common questions from new users is
how to publish raw JSON. It's the most useful mode,
so we should make it easy to use.

Modifications
-------------
Remove the default value for `couchbase.source.handler`.
(I'd love to actually make `RawJsonSourceHandler` the
default, but it requires using a special `type.converter`,
and we don't have control over the default type converter.)

Use `RawJsonSourceHandler` in the quickstart config files.

Update the quickstart guide, and add more documentation
for the various built-in source handlers.

Fix the `tasks.max` property, which was previously changed
accidentally.

Change-Id: I4faee813184ea766f28e95f79ea32022c7336284
Reviewed-on: http://review.couchbase.org/c/kafka-connect-couchbase/+/131477
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>