KAFKAC-226 Change "owner name" of Confluent Hub component archive
Motivation
----------
Confluent unilaterally decided our owner name is "couchbase"
instead of "couchbaseinc". Path of least resistance is for us
to follow suit and rename the archive.

Modifications
-------------
Replace "couchbaseinc" with "couchbase" when referring to the
distribution archive name.

Document that the connector is available on Confluent Hub.

Fix minor typo.

Change-Id: I41a87f69b37bc9d78a010613ae02d89f8ba54a1b
Reviewed-on: http://review.couchbase.org/c/kafka-connect-couchbase/+/138266
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>